got rid of dummy data, implemented a bunch of new stuff

+++ app/Livewire/CVs/Index.php
@@ -3,23 +3,28 @@
 namespace App\Livewire\CVs;
 
 use App\Models\CV;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Index extends Component
 {
-    use WithPagination;
+    use AuthorizesRequests, WithPagination;
 
     public $search = '';
 
     public function delete(CV $cv)
     {
+        $this->authorize('delete', $cv);
+
         $cv->delete();
         session()->flash('message', 'CV deleted successfully.');
     }
 
     public function render()
     {
+        $this->authorize('viewAny', CV::class);
+
         $cvs = CV::with('person')
             ->where('file_path_or_url', 'like', '%'.$this->search.'%')
             ->orWhereHas('person', function ($query) {
